+++ pynyzo/tests/test_ed25519.py
@@ -59,6 +59,7 @@ def test_verify_1(verbose=False):
     except:
         raise
 
+
 if __name__ == "__main__":
     # test_keys_2(True)
     # KeyUtil.main()